test: don't write to stream if test passed successfully
On success case, timeout still tries to fire and write to
stream. Only write to stream if we haven't finished

+++ test/sequential/test-fs-watch-system-limit.js
@@ -37,13 +37,15 @@ function spawnProcesses() {
 spawnProcesses();
 
 setTimeout(() => {
-  const timeoutStream = new stream.PassThrough();
-  timeoutStream.write('Error: timeout');
-  timeoutStream.end();
-  timeoutStream.pipe(gatherStderr);
+  if (!finished) {
+    const timeoutStream = new stream.PassThrough();
+    timeoutStream.write('Error: timeout');
+    timeoutStream.end();
+    timeoutStream.pipe(gatherStderr);
 
-  finished = true;
-  processes.forEach((proc) => proc.kill());
+    finished = true;
+    processes.forEach((proc) => proc.kill());
+  }
 }, timeout);
 
 let accumulated = '';
